// src/shared/types.ts
export interface ServerConfig {
  id: string;
  name: string;
  group: string;
  host: string;
  port: number;
  username: string;
  authMethod: 'password' | 'key' | 'agent';
  password?: string;
  privateKeyPath?: string;
  passphrase?: string;
  keepAlive?: boolean;
  keepAliveInterval?: number;
  description?: string;
  createdAt: number;
  updatedAt: number;
}